animation from pre-render page to current page
- A first version of the UI is generated during the build time with
  Gatsby.
- The background color is set by the current time of the day

The issue is that the pre-rended UI has the background color of the buld
time.

It could be a feature but it's currently a bug 👼.

Solution: I detect when it's pre-render or live render and I generate a
black/white version of the page.

Result: when the page is loaded we're going from a black/white page to a
colorful page with a 100ms animation.